fix prices

+++ PayForMeBot/ReceiptApiClient/Models/Product.cs
@@ -2,9 +2,9 @@ namespace PayForMeBot.ReceiptApiClient.Models;
 
 public class Product
 {
-    public int TotalPrice { get; set; }
+    public double TotalPrice { get; set; }
 
-    public int Price { get; set; }
+    public double Price { get; set; }
 
     public int Count { get; set; }
